+++ app/assets/stylesheets/sessions.scss
