--- app/controllers/private_events_controller.rb
@@ -1,4 +1,0 @@
-class PrivateEventsController < ApplicationController
-  def index
-  end
-end
