--- app/controllers/sessions_controller.rb
@@ -1,16 +1,0 @@
-class SessionsController < ApplicationController
-    
-    def new
-    end
-    
-    def create
-        user = User.find_by_name(params[:name])
-        session[:user_id] = user.id
-    end
-    
-    def destroy
-        session[:user_id] = nil
-        redirect_to root_path
-    end
-
-end
