+++ app/controllers/users_controller.rb
@@ -1,24 +1,25 @@
 class UsersController < ApplicationController
-    include UsersHelper 
-    before_action :require.login, except: %i[new create]
+  def index
+    @users = User.all
+  end
 
-    def index
-        @users = User.all
-    end
-    
-    def new
-        @user= User.new
-    end
-    
-    def create
-        @user = User.new(user_params)
-        @user.save
-        session[:user_id] = user.id
-        redirect_to user_path(@user)
-    end
-    
-    def show
-        @user = User.find(params[:id])
-    end
+  def new
+    @user = User.new
+  end
 
+  def create
+    @user = User.new(user_params)
+    @user.save
+
+    redirect_to user_path(@user)
+  end
+
+  def show
+    @user = User.find(params[:id])
+  end
+
+  Private
+  def user_params
+    params.require(:user).permit(:name)
+  end
 end
