--- app/helpers/private_events_helper.rb
@@ -1,2 +1,0 @@
-module PrivateEventsHelper
-end
