+++ app/helpers/users_helper.rb
@@ -1,5 +1,2 @@
 module UsersHelper
-    def user_params
-        params.require(:user).permit(:name)
-    end
 end
