--- app/models/session.rb
@@ -1,2 +1,0 @@
-class Session < ApplicationRecord
-end
