--- app/views/private_events/index.html.erb
@@ -1,4 +1,0 @@
-<h1>PrivateEvents#index</h1>
-<p>Hola mundo</p>
-
-
