--- app/views/sessions/new.html.erb
@@ -1,1 +1,0 @@
-<h1>New Session</h1>
