--- app/views/users/_form.html.erb
@@ -1,17 +1,0 @@
-<h1>Register New User</h1>
-<%= form_with() do |f| %>
-    <div class="container">
-        <div class="field">
-            <div class="control">
-                <%= f.label :name, class: "label" %>
-                <%= f.text_field :name, class: "input", autofocus: true %>
-            </div>
-        </div>
-
-        <div class="field">
-            <div class="control">
-                <%= f.submit "Sign Up", class: "button is-primary is-large is-fullwidth" %>
-            </div>
-        </div>
-    </div>
-<% end %>
