+++ app/views/users/index.html.erb
@@ -1,1 +1,26 @@
-<h1>Users</h1>
+<nav class="navbar" role="navigation" aria-label="main navigation">
+      <div class="navbar-brand">
+   
+        <%= link_to root_path, class: "navbar-item" do %>
+          <h1 class="title is-4">Private Events</h1>
+        <% end %>
+
+        <a role="button" class="navbar-burger burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample">
+          <span aria-hidden="true"></span>
+          <span aria-hidden="true"></span>
+          <span aria-hidden="true"></span>
+        </a>
+      </div>
+
+      <div id="navbarExample" class="navbar-menu">
+        <div class="navbar-end">
+          <div class="navbar-item">
+            <div class="field is-grouped">
+              <button>Create Event</button>
+              <button>Sign In</button>
+              <button>Sign Up</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </nav>
