--- app/views/users/new.html.erb
@@ -1,2 +1,0 @@
-<h1>Sign Up </h1>
-<%= render 'form', user: @user %>
