+++ app/views/users/show.html.erb
@@ -1,0 +1,1 @@
+<h1><%= @user.name %></h1>
