+++ config/routes.rb
@@ -1,11 +1,8 @@
 Rails.application.routes.draw do
+
+  get 'users/index'
+
+  root to:'users#index'
+
   resources :users
-  resources :sessions, only: [:new, :create, :destroy]
-
-  get 'signup' => 'users#new', as: 'signup'
-  get 'login' => 'sessions#new', as: 'login'
-  get 'logout' => 'sessions#destroy', as: 'logout'
-  
-
-  root to: "users#index"
 end
