--- db/migrate/20210111225303_create_sessions.rb
@@ -1,8 +1,0 @@
-class CreateSessions < ActiveRecord::Migration[6.1]
-  def change
-    create_table :sessions do |t|
-
-      t.timestamps
-    end
-  end
-end
