+++ db/migrate/20210113171200_create_users.rb
@@ -2,7 +2,7 @@ class CreateUsers < ActiveRecord::Migration[6.1]
   def change
     create_table :users do |t|
       t.string :name
-      
+
       t.timestamps
     end
   end
