+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_01_11_225303) do
+ActiveRecord::Schema.define(version: 2021_01_13_171200) do
 
   create_table "sessions", force: :cascade do |t|
     t.datetime "created_at", precision: 6, null: false
