+++ spec/private_events_spec.rb
