+++ spec/requests/users_request_spec.rb
@@ -2,4 +2,11 @@ require 'rails_helper'
 
 RSpec.describe "Users", type: :request do
 
+  describe "GET /index" do
+    it "returns http success" do
+      get "/users/index"
+      expect(response).to have_http_status(:success)
+    end
+  end
+
 end
