+++ spec/views/users/index.html.erb_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Session, type: :model do
+RSpec.describe "users/index.html.erb", type: :view do
   pending "add some examples to (or delete) #{__FILE__}"
 end
